Reverting for loop backend changes

+++ server/src/main/java/com/testsigma/step/processors/WhileLoopStepProcessor.java
@@ -21,7 +21,7 @@ public class WhileLoopStepProcessor extends StepProcessor {
                                 WorkspaceType workspaceType, Map<String, Element> elementMap,
                                 TestStepDTO testStepDTO, Long testPlanId, TestDataSet testDataSet,
                                 Map<String, String> environmentParams, TestCaseEntityDTO testCaseEntityDTO,
-                                String environmentParamSetName, String dataProfile, Map<Long, Long> dataSetIndex) {
+                                String environmentParamSetName, String dataProfile, Map<Long, Integer> dataSetIndex) {
     super(webApplicationContext, testCaseStepEntityDTOS, workspaceType, elementMap, testStepDTO, testPlanId, testDataSet,
       environmentParams, testCaseEntityDTO, environmentParamSetName, dataProfile, dataSetIndex);
   }
